Optimising the get() method for objects

+++ src/document/class/fields.cpp
@@ -404,7 +404,7 @@ static ERR GET_WorkingPath(extDocument *Self, CSTRING *Value)
    bool path = false;
    if (Self->Path[0] IS '/') path = true;
    else {
-     for (LONG j=0; (Self->Path[j]) and (Self->Path[j] != '/') and (Self->Path[j] != '\\'); j++) {
+     for (int j=0; (Self->Path[j]) and (Self->Path[j] != '/') and (Self->Path[j] != '\\'); j++) {
          if (Self->Path[j] IS ':') {
             path = true;
             break;
@@ -412,8 +412,8 @@ static ERR GET_WorkingPath(extDocument *Self, CSTRING *Value)
       }
    }
 
-   LONG j = 0;
-   for (LONG k=0; Self->Path[k]; k++) {
+   int j = 0;
+   for (int k=0; Self->Path[k]; k++) {
       if ((Self->Path[k] IS ':') or (Self->Path[k] IS '/') or (Self->Path[k] IS '\\')) j = k+1;
    }
